To prevent search lag when there are a large number of notes
